# MedWayHealthCare2.0/src/Presentation/ProductAdminPortal/Documentation/API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: ProductAdminPortal API
  description: |
    Domain-agnostic product configuration and subscription management system.
    
    ## Features
    - Multi-tenant product catalog management
    - Flexible pricing engine with tiered, per-unit, and complexity-based pricing
    - Usage tracking and billing calculation
    - Subscription lifecycle management
    - Audit trails and versioning
    - RBAC with granular permissions
    
    ## Authentication
    All endpoints require JWT bearer token authentication.
    
    ## Rate Limiting
    - 1000 requests per hour for authenticated users
    - 100 requests per hour for anonymous users
    
  version: 1.0.0
  contact:
    name: ProductAdminPortal Support
    email: support@productadminportal.com
servers:
  - url: https://api.productadminportal.com/v1
    description: Production server
  - url: https://staging-api.productadminportal.com/v1
    description: Staging server
  - url: http://localhost:5000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        key:
          type: string
        description:
          type: string
        metadata:
          type: object
        version:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Module:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        display_order:
          type: integer
        metadata:
          type: object
        is_active:
          type: boolean
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: List all products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer

    post:
      tags:
        - Products
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                key:
                  type: string
                description:
                  type: string
                metadata:
                  type: object
              required:
                - name
                - key
            example:
              name: "Hospital Management System"
              key: "HMS_CORE"
              description: "Complete hospital management platform"
              metadata:
                industry: "healthcare"
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error
        '409':
          description: Product key already exists
  /products/{productId}:
    get:
      tags:
        - Products
      summary: Get product by ID
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found

    put:
      tags:
        - Products
      summary: Update product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                metadata:
                  type: object
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

    delete:
      tags:
        - Products
      summary: Delete product (soft delete)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Product deleted successfully
  /products/{productId}/modules:
    get:
      tags:
        - Products
        - Modules
      summary: Get modules for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Modules retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Module'

    post:
      tags:
        - Products
        - Modules
      summary: Add module to product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                display_order:
                  type: integer
              required:
                - name
            example:
              name: "Patient Management"
              description: "Complete patient registration and records"
              display_order: 1
      responses:
        '201':
          description: Module added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Module'

  # ===========================
  # ENTITIES
  # ===========================
...